Separating models into their own module.
