+++ server/src/models/domain.js
@@ -1,0 +1,236 @@
+const Base = require('./base');
+const Entity = require('./entity');
+
+//
+// Class "Domain"
+//
+
+// Constructor and inheritance
+
+function Domain (headstart, name, desc, recreate) {
+    // Special case - the parent of domain is headstart, which is not of type "Base"
+    this.id_counter = 1;
+    Base.call(this, "Domain", headstart, this.id_counter, name, desc);
+    this.entities = [];
+    this.definitionOfMany = 100;
+
+    // Create rootEntity entity:
+    if (!recreate) {
+        this.addNewEntity(this.name, "Root for domain " + this.name, null, false, true);
+    }
+}
+Domain.prototype = Object.create(Base.prototype);
+Domain.prototype.constructor = Domain;
+
+// Methods
+
+Domain.prototype.save = function () {
+    var fn = this.getHeadStart().getDir("domains") + this.name + ".jsn";
+    var fs = require('fs');
+    fs.writeFileSync(fn, JSON.stringify(this, null, 2));
+};
+
+Domain.prototype.createNewID = function () {
+    if (this.id_counter < 2) {
+        var max = 1;
+        var all = this.getAllElements();
+        for (var i in all) if (all[i].id > max) max = all[i].id;
+        this.id_counter = max + 1;
+    }
+    return this.id_counter++;
+};
+
+Domain.prototype.compareIDs = function (id1, id2) {
+    return id1.toString() === id2.toString();
+};
+
+Domain.prototype.validateModel = function () {
+    var i;
+    var vRes = "";
+    var wCount = 0;
+
+    if (this.name === "New_Domain") {
+        wCount++;
+        vRes += "<br>[" + wCount + "]: <strong>" + this.name + "</strong> is not a unique name - please rename your Domain!";
+    }
+
+    // All Relationships need targets
+    for (i in this.entities) {
+        for (var j in this.entities[i].relationships) {
+            if (!this.entities[i].relationships[j].hasTargetEntity()) {
+                wCount++;
+                vRes += "<br>[" + wCount + "]: <strong>" + this.entities[i].name + "::" + this.entities[i].relationships[j].name + "</strong> does not have a target!";
+            }
+        }
+    }
+
+    // All entities should either be abstract, root entity or aggregated by another entity (directly or through inheritance):
+    for (i in this.entities) {
+        if (!this.entities[i].isAbstract && !this.entities[i].canBeInstantiated()) {
+            wCount++;
+            vRes += "<br>[" + wCount + "]: <strong>" + this.entities[i].name + "</strong> can not be instantiated (solution: make it a RootEntity or child of another entity)";
+        }
+    }
+
+    if (wCount === 0) return null;
+    return wCount === 1 ? "<strong>1 Warning:</strong>" + vRes : "<strong>" + wCount + " Warnings:</strong>" + vRes;
+};
+
+Domain.prototype.getEntities = function (sortByInheritance) {
+    if (!sortByInheritance) { return this.entities; }
+    for (var i in this.entities) {
+        var entity = this.entities[i];
+        entity.longName = entity.name;
+        var tmpEntity = entity;
+        while (tmpEntity.hasParentClass()) {
+            tmpEntity = tmpEntity.getParentClass();
+            entity.longName = tmpEntity.name + ":" + entity.longName;
+        }
+    }
+    return this.entities.sort(function (a, b) {
+        if (a.longName < b.longName) return -1;
+        if (a.longName > b.longName) return 1;
+        return 0;
+    });
+};
+
+Domain.prototype.getRootEntities = function () {
+    var allEntities = this.entities;
+    var rootEntities = [];
+    for (var i in allEntities) {
+        if (allEntities[i].isRootEntity) { rootEntities.push(allEntities[i]); }
+    }
+    return allEntities;
+};
+
+Domain.prototype.getRootInstance = function () {
+    var allEntities = this.entities;
+    for (var i in allEntities) {
+        if (allEntities[i].isRootInstance) { return allEntities[i]; }
+    }
+    throw new Error("Domain without root instance!");
+};
+
+Domain.prototype.addEntity = function (newEntity) {
+    return this.entities.push(newEntity);
+};
+
+Domain.prototype.addNewEntity = function (name, desc, parentClass, isRootEntity, isRootInstance) {
+    var id = this.getDomain().createNewID();
+    var newEntity = new Entity(this, id, name, desc, parentClass, isRootEntity, isRootInstance);
+
+    this.addEntity(newEntity);
+    return newEntity;
+};
+
+Domain.prototype.createNewDefaultViews = function () {
+    this.getEntities().forEach(function (elem) {
+        elem.createNewDefaultViews();
+    });
+};
+
+Domain.prototype.getAllElements = function (includeSelf) {
+    // Returns an array containing all child elements; optional: also include self
+    var r = [];
+    if (includeSelf) { r = r.concat(this); }
+    for (var i in this.getEntities()) { r = r.concat(this.getEntities()[i].getAllElements(true)); }
+    return r;
+};
+
+Domain.prototype.createTestDataForEntity = function (entityDef, relationship, parentInstanceID, parentBaseClass) {
+    // TBD1: Change algorithm to  create as many entities as possible with one DB insert
+    // TBD2: Leverage cardinality to "embedd" child objects with low cardinality into the parent document
+
+    if (entityDef.isAbstract) { return; } // Don't create test instances for abstract entity types!
+    var testData = entityDef.createTestInstance();
+
+    if (parentInstanceID) {
+        testData.parentID = parentInstanceID;
+        testData.parentRelnID = relationship.id;
+        testData.parentRelnName = relationship.name;
+        testData.parentBaseClassID = parentBaseClass.id;
+        testData.parentBaseClassName = parentBaseClass.name;
+    } else {
+        testData.isRootInstance = true;
+        testData.parentID = null;
+        testData.parentRelID = null;
+        testData.parentBaseClassID = null;
+        testData.parentBaseClassName = null;
+    }
+    // TBD: TEST - var ObjectID = require("mongodb").ObjectID;
+    // testData._id = new ObjectID().toString();
+
+    var domain = this;
+    this.getHeadStart().useDB(domain.name, function (db) {
+        var collection = db.collection(entityDef.getBaseClass().name);
+        collection.insertOne(testData, function (mongoErr, mongoRes) {
+            if (mongoErr) {
+                console.log("Error creating test data: " + mongoErr);
+            } else {
+                var aggs = entityDef.getAggregations();
+                if (aggs) {
+                    aggs.forEach(function (rel) {
+                        var avg = rel.targetAverage;
+                        if (avg === "/") { console.log("WARNING: Incomplete Quantity Model - Average for relationship '" + rel.name + "' not defined!"); }
+                        for (var i = 0; i < avg; i++) { domain.createTestDataForEntity(rel.getTargetEntity(), rel, mongoRes.ops[0]._id, entityDef.getBaseClass()); }
+                    });
+                }
+            }
+        });
+    });
+};
+
+Domain.prototype.processLocalTemplateFunctions = function (template) {
+    var children = [["Entity", this.getEntities(true)]];
+    template = this.processTemplateWithChildElements(template, children);
+    return Base.prototype.processLocalTemplateFunctions.call(this, template);
+};
+
+Domain.prototype.toJSON = function () {
+    var e = [];
+    for (var i in this.getEntities()) e.push(this.getEntities()[i].toJSON());
+    return {
+        name: this.name,
+        desc: this.desc,
+        type: this.type,
+        id: this.idToJSON(),
+        definitionOfMany: this.definitionOfMany,
+        entities: e
+    };
+};
+
+Domain.prototype.toString = function () {
+    var e;
+    var es;
+    var i;
+    var s = "//\n// Domain '" + this.name + "'\n//\n";
+
+    s += "\n// Basic Entity Definitions:\n";
+    for (i in this.getEntities()) {
+        e = this.getEntities()[i];
+        s += e.toString("properties") + "\n";
+    }
+
+    s += "\n// Aggregation Hierarchy:\n";
+    for (i in this.getEntities()) {
+        e = this.getEntities()[i];
+        es = e.toString("aggregations");
+        if (es.length > 0) { s += es + "\n"; }
+    }
+
+    s += "\n// Associations:\n";
+    for (i in this.getEntities()) {
+        e = this.getEntities()[i];
+        es = e.toString("associations");
+        if (es.length > 0) { s += es + "\n"; }
+    }
+
+    return s;
+};
+
+Domain.prototype.getFileName = function (folder) {
+    var f = folder ? folder + "\\" : "";
+    var fn = '.\\generated\\' + f + this.name + '.html';
+    return fn;
+};
+
