+++ server/src/models/index.js
@@ -1,0 +1,17 @@
+const Base = require('./base');
+const BasicProperty = require('./basic-property');
+const Entity = require('./entity');
+const Enumeration = require('./enumeration');
+const Literal = require('./literal');
+const Relationship = require('./relationship');
+const views = require('./views');
+
+module.exports = {
+    Base,
+    Literal,
+    Enumeration,
+    BasicProperty,
+    Relationship,
+    Entity,
+    views
+};
