+++ server/src/models/views.js
@@ -1,0 +1,21 @@
+const BasicPropertyPanelItem = require('./basic-property-panel-item');
+const EnumPanelItem = require('./enum-panel-item');
+const PageView = require('./page-view');
+const Panel = require('./panel');
+const PanelItem = require('./panel-item');
+const RelationshipPanelItem = require('./relationship-panel-item');
+const SeparatorPanelItem = require('./separator-panel-item');
+const TableItem = require('./table-item');
+const TableView = require('./table-view');
+
+module.exports = {
+    BasicPropertyPanelItem,
+    EnumPanelItem,
+    PageView,
+    Panel,
+    PanelItem,
+    RelationshipPanelItem,
+    SeparatorPanelItem,
+    TableItem,
+    TableView
+};
